+++ angular/index.controllers.js
@@ -1,4 +1,0 @@
-import { LandingController } from './app/pages/landing/landing.controller';
-
-angular.module('app.controllers')
-	.controller('LandingController', LandingController);