+++ angular/index.main.js
@@ -4,9 +4,6 @@ import './index.modules';
 //Config
 import './index.config';
 
-//Controllers
-import './index.controllers';
-
 //Run
 import './index.routes';