+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - reverted to `gulp-minify-css` [#144](https://github.com/jadjoubran/laravel5-angular-material-starter/issues/144)
 - EcmaScript 6 generators thanks to [@jsalhani](https://github.com/jsalhani) [#152](https://github.com/jadjoubran/laravel5-angular-material-starter/pull/152)
 - Ignore (stop tracking) public/js and public/css
+- Remove all controllers to promote components (Angular 1.5) [#154](https://github.com/jadjoubran/laravel5-angular-material-starter/issues/154)
 
 ### 3.0.6
 - npm dependencies cleanup [#133](https://github.com/jadjoubran/laravel5-angular-material-starter/issues/133)